+++ BootCamp/BootCamp/Pages/LoginPage.cs
@@ -1,4 +1,5 @@
-﻿using BootCamp.Pages.Base;
+﻿using BootCamp.Model;
+using BootCamp.Pages.Base;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
 using OpenQA.Selenium;
 using System;
@@ -11,10 +12,11 @@ namespace BootCamp.Pages
 {
     public class LoginPage : TestShopPage
     {
-
         private By loginEmailField = By.CssSelector("input#email");
         private By loginPasswordField = By.CssSelector("input#passwd");
         private By loginSubmitButton = By.CssSelector("button#SubmitLogin");
+        private By signInEmailField = By.CssSelector("input#email_create");
+        private By signInSubmitButton = By.CssSelector("button#SubmitCreate");
 
         public LoginPage(IWebDriver driver) : base(driver)
         {
@@ -29,6 +31,13 @@ namespace BootCamp.Pages
             Assert.AreEqual("MY ACCOUNT", GetPageTitle());
 
             return new MyAccountPage(driver);
+        }
+
+        public MyAccountPage CreateNewUserAccount(UserAccount user)
+        {
+            driver.FindElement(signInEmailField).SendKeys(user.email);
+            driver.FindElement(signInSubmitButton).Click();
+            return new NewAccountPage(driver).CreateNewUserAccount(user);
         }
     }
 }
