+++ BootCamp/BootCamp/Test/Base/TestShopScenario.cs
@@ -30,7 +30,7 @@ namespace BootCamp.Test.Base
         [TestCleanup]
         public void TearDown()
         {
-            //driver.Quit();
+            driver.Quit();
         }
     }
 }
